requirements.txt
Django==3.2.25
django-filter==22.1
requests==2.31.0
feedparser==6.0.11
pandas==1.3.5
plotly==5.18.0
dj-database-url==0.5.0
pytest==7.1.3
pytest-django==4.5.2
factory-boy==3.2.1
coverage==6.5.0
black==22.3.0
ruff==0.0.285
django-distill==3.2.0
